refactor: rename package comply_with_me/ → cwm/
The entry point script (comply_with_me.py) and the application package
shared the same base name, causing visual confusion and potential import
ambiguity. The script is the user-facing entry point and keeps its name;
the package is an implementation detail and is renamed to cwm.

- git mv comply_with_me/ cwm/
- Update all from comply_with_me.xxx imports → from cwm.xxx across the
  package and the bootstrap script
- pyproject.toml: update [project.scripts] entry point and packages.find
  include pattern to cwm*
- cmmc.py: update user-facing notice path reference to cwm/downloaders/cmmc.py
