+++ cwm/state.py
